Case category constants for display
Ultimately, these will be replaced by a proper ProgramCategory model. In the meantime, correcting the casing here does away with the need to force consumers of this API to parse and hardcode references to these categories. Django doesn't enforce choices on a database level; the migration exists to appease Django 1.8's migrations framework. Part of ECOM-5018.